bot/requirements.txt
# Core Discord bot dependencies
discord.py>=2.3.0
aiohttp>=3.8.0

# Environment and configuration
python-dotenv>=0.19.0

# Language detection
langdetect>=1.0.9

# Database support
aiosqlite>=0.19.0
asyncpg>=0.28.0

# Caching (optional)
redis>=4.5.0

# System monitoring (optional)
psutil>=5.9.0

# HTTP requests
requests>=2.25.0

# Async utilities
asyncio-throttle>=1.0.0

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0